+++ Classes/CDI.h
@@ -21,13 +21,16 @@
 //    CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
 
+#import <Foundation/Foundation.h>
+#import "CDISingleton.h"
+#import "CDIInterceptor.h"
 
 // This prefix is used to identify the injectable instance variable
 #define __INJECT_INSTANCE_PREFIX @"__inject_instance__"
 #define __INJECT_TYPE_PREFIX @"__inject_type__"
 
 /**
- * The inject macro which is can used to create an instance and assign
+ * The inject macro which is used to create an instance and assign
  * it to the instance variable of a class.
  *
  * Use:
@@ -36,29 +39,6 @@
  */
 #define inject(instanceVariableName,...) synthesize instanceVariableName;\
 -(void) __inject_instance__ ## __VA_ARGS__ ## __inject_type__ ## instanceVariableName {}
-
-/**
- * The singleton macro generates the sharedInstance method automatically. It creates
- * a new category with the name CDISingleton for the implementation.
- *
- * Use:
- * @singleton(name)
- */
-#define singleton(name) interface name (ContextAndDependencyInjection)\
-+ (instancetype)sharedInstance;\
-@end\
-@implementation name (ContextAndDependencyInjection)\
-+ (instancetype)sharedInstance\
-{\
-    static id sharedInstance = nil;\
-    static dispatch_once_t onceToken;\
-    dispatch_once(&onceToken, ^{\
-        sharedInstance = [[self alloc] init];\
-    });\
-    return sharedInstance;\
-}\
-@end\
-
 
 /**
  * CDIException are thrown by the cdi implementation whenever a unrecoverable situation occurs.