+++ Classes/CDI.m
@@ -61,6 +61,11 @@
 }
 
 - (void)createInstance:(NSString *)variableName inObject:(id)object ofType:(NSString *)type {
+    
+    Ivar ivar = class_getInstanceVariable([object class], [variableName UTF8String]);
+    // If it already is set, do not override it
+    if(!object_getIvar(object, ivar)) {
+    
     // Create the instance and assign it to the instance variable
     id instanceVariableClassType = [self getClassForType:type];
 
@@ -78,8 +83,9 @@
         // Use default instance creation with alloc and init
         instanceVariableValue = [[instanceVariableClassType alloc] init];
     }
-    Ivar ivar = class_getInstanceVariable([object class], [variableName UTF8String]);
-    object_setIvar(object, ivar, instanceVariableValue);
+        object_setIvar(object, ivar, instanceVariableValue);
+    }
+
 }
 
 - (void)bindClass:(Class)instanceVaiableClassType with:(Class)implementationClass {
@@ -248,12 +254,16 @@
 #pragma mark Objective-C Integration
 
 /**
- * The doInject methode is called before the init on any other NSObject creation. It will replace the
+ * The doInject method is called before the init on any other NSObject creation. It will replace the
  * @inject annotation with an instance.
  */
 #pragma clang diagnostic push
 #pragma ide diagnostic ignored "InfiniteRecursion"
 - (id)initWithInject {
+    
+    id returnSelf =nil;
+    
+    NSMutableArray *interceptors = nil;
     
     // Avoid unlimited loop - injection will not be available in the CDI implementation itself
     if ([self isKindOfClass:[CDI class]]) {
@@ -269,6 +279,8 @@
             
             NSString *methodName = [NSString stringWithUTF8String:sel_getName(method_getName((methods[methodIndex])))];
             
+            // Check for instance variable injection
+            // -------------------------------------
             if ([methodName hasPrefix:__INJECT_INSTANCE_PREFIX]) {
                 // Remove the prefix and use it as the instance variable
                 NSString *instanceVariableName = [methodName substringFromIndex:[__INJECT_INSTANCE_PREFIX length]];
@@ -286,12 +298,88 @@
                     [[CDI sharedInstance] createInstance:instanceVariableName inObject:self ofType:type];
                 }
             }
+            
+            // Check for interception injection
+            // -------------------------------------
+            if ([methodName hasPrefix:__INJECT_INTERCEPTOR]) {
+                
+                if(!interceptors) interceptors = [[NSMutableArray alloc] init];
+                    // Get the interceptor class name which is encoded in the method name
+                NSString *interceptorClassName = [methodName substringFromIndex:[__INJECT_INTERCEPTOR length]];
+                
+                    id interceptorClass = [[CDI sharedInstance] getClassForType:interceptorClassName];
+                    // Create instance
+                    id interceptorInstance = [[interceptorClass alloc] init];
+                
+                    // [interceptors insertObject:interceptorInstance atIndex:[interceptors count]];
+                [interceptors insertObject:interceptorInstance atIndex:0];
+                
+                    // [[CDI sharedInstance] injectInterceptor:interceptorClassName inInstance:self];
+            }
         }
         free((void *) methods);
     }
     // Now call the init (remember init was renamed to doInject)
-    return [self initWithInject];
+    if(!returnSelf) returnSelf = [self initWithInject];
+    
+    // Process the interceptor injections
+    if(interceptors) {
+        returnSelf = [[CDIInterceptorProxy alloc] initWithTarget:returnSelf];
+        [returnSelf setInterceptors:interceptors];
+    }
+    
+    return returnSelf;
 }
+
+/**
+ * Inject the interceptor into the instance.
+ */
+-(void)injectInterceptor:(NSString*)interceptorName inInstance:(id)instance {
+    
+    id interceptorClass = [self getClassForType:interceptorName];
+    // Create instance
+    id interceptorInstance = [[interceptorClass alloc] init];
+    
+    unsigned int methodIndex = 0;
+    
+    Method *methods = class_copyMethodList([instance class], &methodIndex);
+    
+    if (methods) {
+        while (methodIndex--) {
+            
+            NSString *methodName = [NSString stringWithUTF8String:sel_getName(method_getName((methods[methodIndex])))];
+            
+            if ([methodName hasPrefix:__INJECT_INTERCEPTOR] || [methodName hasPrefix:__INJECT_METHOD]) continue;
+            
+            Method originalMethod = methods[methodIndex];
+            SEL interceptorSelector = sel_registerName([[NSString stringWithFormat:@"%@__%@__%@",__INJECT_METHOD, interceptorName, methodName] UTF8String]);
+            
+            // The new implementation which will be used to replace the original method
+            IMP newImplementationForMethod = imp_implementationWithBlock(^(id self, SEL releaseSelector, ...) {
+                
+                    // NSLog(@">>> %@",releaseSelector);
+                
+//                CDIInvocationContext *invocationContext = [[CDIInvocationContext alloc] initWithTarget:instance methodName:methodName implementationSelector:interceptorSelector];
+            
+                
+                    // method_setImplementation(class_getClassMethod([interceptorInstance class], @selector(execute)),newIMP);
+                
+                    // id returnValue = [interceptorInstance invoke:invocationContext];
+                
+                return nil;
+            });
+            
+            // Inject the interception method to the class and set the implementation of the
+            // original method.
+            if (class_addMethod([instance class], interceptorSelector, method_getImplementation(originalMethod), method_getTypeEncoding(originalMethod))) {
+                // Replace the implementation of the original method with new implementation
+                method_setImplementation(originalMethod,newImplementationForMethod);
+            }
+        }
+        free((void *) methods);
+    }
+}
+
 #pragma clang diagnostic pop
 
 /**
@@ -307,13 +395,31 @@
     }
 }
 
++(instancetype)allocWithInitInjection {
+    return [self allocWithInitInjection];
+}
+
+/**
+ * Swap the implementation of the [NSObject init] with [CDI doInject].
+ */
++(void) swapAllocWithAllocInjectMethod {
+    Method allocMethod = class_getClassMethod([NSObject class], @selector(alloc));
+    Method allocWithInitInjectionMethod = class_getClassMethod([CDI class], @selector(allocWithInitInjection));
+    
+    if (class_addMethod([NSObject class], @selector(allocWithInitInjection), method_getImplementation(allocMethod), method_getTypeEncoding(allocMethod))) {
+        method_exchangeImplementations(allocMethod, allocWithInitInjectionMethod);
+    }
+}
+
 /**
  * Initialize will replace the init implementation of NSObject.
  */
 + (void) initialize {
     if (self == [CDI class]) {
+        [self swapAllocWithAllocInjectMethod];
             [self swapInitWithDoInjectMethod];
     }
 }
+
 
 @end