+++ Project/Sample/Sample/MySample1ServiceImplemetation.h
@@ -20,14 +20,10 @@
 //    IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 //    CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
-#import <UIKit/UIKit.h>
-#import "MySampleService.h"
 
-@interface ViewController : UIViewController {
-}
-@property(nonatomic) id<MySampleService> sampleService;
-@property(nonatomic) NSString *status;
-@property(nonatomic) NSNumber *count;
-@property (weak, nonatomic) IBOutlet UITextView *logArea;
-- (IBAction)doIt:(id)sender;
+#import <Foundation/Foundation.h>
+#import "MySample1Service.h"
+
+@interface MySample1ServiceImplemetation : NSObject <MySample1Service>
+@property(nonatomic,readonly) NSNumber *count;
 @end